requirements.txt
asgiref==3.5.2
Django==3.2
django-cors-headers==3.10.1
django-extensions==3.1.5
djangorestframework==3.13.1
Pillow==8.0.1
pytz==2022.1
sqlparse==0.4.2
